+++ CHANGES.rst
@@ -1,8 +1,26 @@
 Changes
--------
+=======
+
+0.4 (unreleased)
+------------------
+
+- Removed old code and templates
+  [djay]
+- Added buildout and robot tests for Plone 4.1, 4.2, 4.3
+  [sgeulette]
+- Dont't skip password validation for manager
+  [sgeulette]
+- Skip password validation for generated password
+  [sgeulette]
+- Added i18n and french translation.
+  [sgeulette]
+- Added travis configuration
+  [sgeulette]
+- Added pwreset_form and test
+  [sgeulette]
 
 0.3.2 (unreleased)
-==================
+------------------
 
 - Updated README file. [macagua]
 - Added QA and testing buildout configuration. [macagua]
@@ -11,22 +29,22 @@ Changes
 - Added i18n support. [macagua]
 
 0.3.1 (2013-11-20)
-==================
+------------------
 
 - Bugfix for use inside change-password
 - Don't validate password strength of old password
-  (pysailor)
+  [pysailor]
 
 0.3 (2013-08-18)
-================
+----------------
 
-* Added a monkey-patch for the zope.schema Password field to validate
+- Added a monkey-patch for the zope.schema Password field to validate
   the password. This is necessary for the Plone 4 @@new-user form to
   work well. [regebro]
 
 
 Earlier versions
-================
+----------------
 
 0.2 - Packaged as egg. Plone 3.1 compatible